requirements.txt
asgiref==3.4.1
boto3==1.20.50
botocore==1.23.50
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.7
django-dbbackup==4.0.1
django-anymail==8.5
django-cors-headers==3.12.0
django-filter==21.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pre-commit
psycopg2-binary==2.9.3
pycparser==2.20
pyngrok==5.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
